requirements.txt
# AI News Intelligence Dashboard — Python dependencies
# Install: pip install -r requirements.txt

# --- Core ---
streamlit>=1.30.0
pandas>=2.0.0
requests>=2.31.0
feedparser>=6.0.0
python-dotenv>=1.0.0

# --- NLP & ML ---
scikit-learn>=1.3.0
nltk>=3.8.0

# --- Visualisation ---
plotly>=5.18.0

# --- Testing ---
pytest>=7.4.0
pytest-cov>=4.1.0

# --- Optional AI layer (uncomment if you want LLM summaries) ---
# openai>=1.12.0
# anthropic>=0.18.0
